requirements.txt
torch>=1.8.0
transformers>=4.5.0
pandas>=1.2.0
numpy>=1.19.0
tqdm>=4.50.0
jieba>=0.42.0
matplotlib>=3.3.0
seaborn>=0.11.0
tensorboard>=2.4.0
scikit-learn>=0.24.0
tldextract>=3.1.0
wordcloud>=1.8.0
pillow>=8.0.0 